html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-move-source-into-empty-video.html

<!doctype html>
<title>pointer updates (moving a source element into an empty media element)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<div id="holder"></div>
<script>
var events = [];
var b = 0;
var x = 0;
</script>
<video id="v"><source id="source-b" onerror="b++; events.push('b')" src="resources/delayed-broken-video.py"></video>
<script>
async_test(function(t) {
  var video = document.getElementById('v');
  var holder = document.getElementById('holder');

  // Remove the current candidate so the video has no children while b is pending.
  holder.appendChild(document.getElementById('source-b'));

  var sourceX = document.createElement('source');
  sourceX.onerror = function() { x++; events.push('x'); };
  holder.appendChild(sourceX);

  // Move an existing source from another parent into the empty video.
  video.appendChild(sourceX);

  window.onload = t.step_func(function() {
    assert_equals(b, 1, 'error events on b');
    assert_equals(x, 1, 'error events on x');
    assert_array_equals(events, ['b', 'x'], 'processed source order');
    t.done();
  });
});
</script>
